Moving CommandLineParser and factory methods out
...to a surrogate main package
